requirements.txt
requests>=2.31.0
actualpy>=0.1.0
python-dotenv>=1.0.0
